+++ src/main/java/gift/service/ProductService.java
@@ -1,23 +1,20 @@
 package gift.service;
 
 import gift.domain.Product;
-import gift.domain.ProductDaoImpl;
+import gift.dao.ProductDao;
 import gift.exception.DuplicateProductNameException;
-import gift.domain.ProductValidator;
-
 import java.util.List;
 import org.springframework.stereotype.Service;
 
 @Service
 public class ProductService {
 
-    private final ProductDaoImpl productDao;
+    private final ProductDao productDao;
 
 
-    private final ProductValidator productValidator = new ProductValidator();
 
 
-    public ProductService(ProductDaoImpl productDao) {
+    public ProductService(ProductDao productDao) {
         this.productDao = productDao;
         productDao.createProductTable();
     }
